# order/docs/swagger.yaml
definitions:
  order.Request:
    properties:
      id:
        type: string
      products:
        items:
          type: string
        type: array
      status:
        type: string
      total_price:
        type: number
      user_id:
        type: string
    type: object
  order.Response:
    properties:
      id:
        type: string
      products:
        items:
          type: string
        type: array
      status:
        type: string
      total_price:
        type: number
      user_id:
        type: string
    type: object
  response.Object:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /orders:
    get:
      consumes:
      - application/json
      description: Get all orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/order.Response'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Add a new order
      parameters:
      - description: Order request
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Add an order
      tags:
      - orders
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order deleted
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Delete an order
      tags:
      - orders
    get:
      consumes:
      - application/json
      description: Get order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Get an order
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Update order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Order request
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order.Request'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Object'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Update an order
      tags:
      - orders
  /orders/search:
    get:
      consumes:
      - application/json
      description: Search orders by user ID or status
      parameters:
      - description: User ID
        in: query
        name: userId
        type: string
      - description: Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/order.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Search orders
      tags:
      - orders
swagger: "2.0"